Small modification to repo URL to enable cloning vpngw again.
The old URL to clone vpngw stopped working for the last couple of days with
an error saying:
fatal: could not read Username for 'https://github.com': No such device or address
The main SCION codebase can still be fetched the existing way. I suspect
the difference might be due to the vpngw repo being private.